move Rare component to goblins
